Add MP4A support for macOS with minimp4 and AVFoundation
This PR add initial mp4a support for macOS with minimp4
for container parsing and macOS's native AVFoundation in swift
for decoding.

The swift code for calling AVFoundation is exported as __cdecl
so that it could be used by C++/C.

This only works on macOS.
